refactor: rename transforms/render.py to transforms/bitmap.py
