+++ tests/transforms/bitmap/_helpers.py
