+++ tests/transforms/bitmap/test_render_bitmap.py
