+++ torchfont/transforms/bitmap.py
